refactor : Result 추상 클래스 구현하여 MoneyResult와 SimpleResult 리팩토링
